Task 1
Currently, a list of products is declared inside the onCreate function of MainActivity.kt, however, it may be used from any part of the app. Your task is to:

create a singleton object ProductsWarehouse.kt
move the list declaration from onCreate to ProductsWarehouse.kt
update the usage of productsList
When a user clicks on a product image the app should open a new activity with its details. The activity and UI are already implemented in the ProductActivity.kt file.